requirements.txt
-f /usr/share/pip-wheels
appdirs==1.4.4
asgiref==3.3.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
certifi==2020.11.8
cffi==1.14.5
chardet==3.0.4
click==7.1.2
contextlib2==0.6.0.post1
crispy-tailwind==0.5.0
cryptography==3.4.6
dateparser==1.0.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
Django==3.1.3
django-admin-interface==0.14.1
django-allauth==0.44.0
django-annoying==0.10.6
django-base64field==1.0
django-ckeditor==6.1.0
django-colorfield==0.3.2
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-datetime-widget==0.9.3
django-email-verification==0.1.0
django-faicon==0.1.0
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-friendship==0.8.0
django-import-export==2.4.0
django-jet==1.0.8
django-js-asset==1.2.2
django-jsignature==0.10
django-material==1.7.3
django-resized==0.3.11
django-simple-history==2.12.0
django-social-auth==0.7.28
django-social-share==2.1.0
django-summernote==0.8.11.6
django-user-unique-email==0.1.1
Django-Verify-Email==0.0.5
djangorestframework==3.12.2
docopt==0.6.2
et-xmlfile==1.0.1
filelock==3.0.12
gmane==0.1.dev25
google-search==1.1.1
greenlet==1.1.1
h11==0.12.0
httpcore==0.13.2
httplib2==0.19.0
httpx==0.18.1
idna==2.10
jdcal==1.4.1
jet-bridge-base==1.0.2
jet-django==1.1.5
lxml==4.6.2
mailchimp-marketing==3.0.58
Markdown==3.3.3
MarkupPy==1.14
mysqlclient==2.0.3
networkx==2.5
newsapi-python==0.2.6
oauth2==1.9.0.post1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
packaging==20.9
Pillow==8.0.1
prompt-toolkit==2.0.9
pycparser==2.20
PyJWT==2.0.1
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-openid==2.2.5
python-slugify==5.0.0
python-social-auth==0.3.6
python3-openid==3.2.0
pythonanywhere==0.9.6
pytube==10.7.2
pytube3==9.6.4
pytz==2020.4
PyYAML==5.3.1
regex==2021.8.28
requests==2.25.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
schema==0.7.4
selenium==3.141.0
six==1.15.0
sniffio==1.2.0
social==0.1.dev3
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
splinter==0.14.0
SQLAlchemy==1.4.23
sqlparse==0.4.1
tablib==2.0.0
tabulate==0.8.9
text-unidecode==1.3
typer==0.3.2
typing-extensions==3.7.4.3
tzlocal==3.0
urllib3==1.26.1
virtualenv==20.2.2
wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0
youtube-search==1.1.1
youtube-search-python==1.4.5
